--- src/main/java/net/k3nder/gl/graphic/sdd
@@ -1,1 +1,0 @@
-asdasd
